+++ lib/providers/data_provider.dart
@@ -6,13 +6,19 @@ import 'package:flutter_fit_utils_provider/invalid_object.dart';
 
 /// Provider containing a single [Modelable] object.
 abstract class DataProvider<T extends Modelable> extends FitProvider {
-  final Service<T> _service;
-
-  /// Returns the service of the provider.
-  Service<T> getService() => _service;
+  /// Service of the provider.
+  late final Service<T> service;
 
   /// Will create [T] for the user if none is found.
   final bool createIfDontExist;
+
+  /// If set to [True], when creating a new instance, it will automatically be
+  /// assigned [userId].
+  final bool assignUserIdOnCreate;
+
+  /// If set to [True], when updating a new instance, it will automatically be
+  /// assigned [userId].
+  final bool assignUserIdOnUpdate;
 
   /// Factory function to create an instance of [T].
   final T Function() factoryFunc;
@@ -37,8 +43,17 @@ abstract class DataProvider<T extends Modelable> extends FitProvider {
   }
 
   /// Creates a new [SingleDataProvider].
-  DataProvider(this._service, this.factoryFunc,
-      {this.createIfDontExist = true});
+  DataProvider(
+    Service<T>? service,
+    this.factoryFunc, {
+    this.createIfDontExist = true,
+    this.assignUserIdOnCreate = true,
+    this.assignUserIdOnUpdate = true,
+  }) {
+    if (service != null) {
+      this.service = service;
+    }
+  }
 
   @override
   Future<void> initialize({dynamic data, String? userId = ""}) async {
@@ -46,14 +61,16 @@ abstract class DataProvider<T extends Modelable> extends FitProvider {
       return;
     }
 
+    await service.repository.initialize();
+
     this.userId = userId ?? "";
 
-    final allData = await _service.getAll(userId: userId);
+    final allData = await service.getAll(userId: userId);
     if (allData.isNotEmpty) {
       _data = allData.first;
     } else if (createIfDontExist) {
       T newData = factoryFunc().copyWith(userId: userId) as T;
-      newData = newData.copyWith(id: await _service.create(newData)) as T;
+      newData = newData.copyWith(id: await service.create(newData)) as T;
 
       _data = newData;
     } else {
@@ -73,8 +90,11 @@ abstract class DataProvider<T extends Modelable> extends FitProvider {
       return (false, null);
     }
 
-    newData = newData.copyWith(userId: userId) as T;
-    _data = newData.copyWith(id: await _service.create(newData));
+    if (assignUserIdOnCreate) {
+      newData = newData.copyWith(userId: userId) as T;
+    }
+
+    _data = newData.copyWith(id: await service.create(newData));
 
     notifyListeners();
 
@@ -90,8 +110,11 @@ abstract class DataProvider<T extends Modelable> extends FitProvider {
       return false;
     }
 
-    _data = _data.copyWith(userId: userId);
-    await _service.update(_data as T);
+    if (assignUserIdOnUpdate) {
+      _data = _data.copyWith(userId: userId);
+    }
+
+    await service.update(_data as T);
 
     notifyListeners();
 
@@ -105,7 +128,7 @@ abstract class DataProvider<T extends Modelable> extends FitProvider {
       return false;
     }
 
-    await _service.delete(_data as T);
+    await service.delete(_data as T);
     notifyListeners();
 
     return true;