+++ lib/providers/items_provider.dart
@@ -4,10 +4,16 @@ import '../flutter_fit_utils_provider.dart';
 
 /// Provider that contains a list of [Modelable] objects.
 abstract class ItemsProvider<T extends Modelable> extends FitProvider {
-  final Service<T> _service;
+  /// Service of the provider.
+  late final Service<T> service;
 
-  /// Returns the service of the provider.
-  Service<T> getService() => _service;
+  /// If set to [True], when creating a new instance, it will automatically be
+  /// assigned [userId].
+  final bool assignUserIdOnCreate;
+
+  /// If set to [True], when updating a new instance, it will automatically be
+  /// assigned [userId].
+  final bool assignUserIdOnUpdate;
 
   /// Factory function to create an instance of [T].
   final T Function() factoryFunc;
@@ -23,7 +29,16 @@ abstract class ItemsProvider<T extends Modelable> extends FitProvider {
   }
 
   /// Creates a new [ItemsProvider].
-  ItemsProvider(this._service, this.factoryFunc);
+  ItemsProvider(
+    Service<T>? service,
+    this.factoryFunc, {
+    this.assignUserIdOnCreate = true,
+    this.assignUserIdOnUpdate = true,
+  }) {
+    if (service != null) {
+      this.service = service;
+    }
+  }
 
   @override
   Future<void> initialize({dynamic data, String? userId = ""}) async {
@@ -31,8 +46,11 @@ abstract class ItemsProvider<T extends Modelable> extends FitProvider {
       return;
     }
 
+    await service.repository.initialize();
+
     this.userId = userId ?? "";
-    _data = await _service.getAll(userId: userId);
+
+    _data = await service.getAll(userId: userId);
 
     initialized = true;
   }
@@ -47,9 +65,12 @@ abstract class ItemsProvider<T extends Modelable> extends FitProvider {
       return (false, null);
     }
 
-    newData = newData.copyWith(userId: userId) as T;
+    if (assignUserIdOnCreate) {
+      newData = newData.copyWith(userId: userId) as T;
+    }
+
     _data.add(
-        newData = newData.copyWith(id: await _service.create(newData)) as T);
+        newData = newData.copyWith(id: await service.create(newData)) as T);
 
     notifyListeners();
 
@@ -66,10 +87,13 @@ abstract class ItemsProvider<T extends Modelable> extends FitProvider {
       return false;
     }
 
-    existingData = existingData.copyWith(userId: userId) as T;
+    if (assignUserIdOnUpdate) {
+      existingData = existingData.copyWith(userId: userId) as T;
+    }
+
     _data.replace(existingData);
 
-    await _service.update(existingData);
+    await service.update(existingData);
 
     notifyListeners();
 
@@ -85,7 +109,7 @@ abstract class ItemsProvider<T extends Modelable> extends FitProvider {
     }
 
     _data.remove(toDelete);
-    await _service.delete(_data as T);
+    await service.delete(_data as T);
 
     notifyListeners();