+++ CHANGELOG.md
@@ -11,6 +11,18 @@
 - Added a getter for _service inside DataProvider
 - Added a getter for _service inside ItemsProvider
 
-## 1.1.0
+## 1.0.2
 
+- Added service initialization inside the initialization of DataProvider and ItemsProvider
+
+## 1.0.3
+
+- Added option to not automatically assign the userId to an instance in DataProvider and ItemsProvider.
+
+## 2.0.0
+
+- Made service public and late initializable in DataProvider and ItemsProvider.
+- Removed getService() in DataProvider and ItemsProvider.
+
+## 2.1.0
 - Made userId parameter nullable in DataProvider.initialize() and ItemsProvider.initialize()